Move provider to new "provider" package
This is done to allow testing only part of the resources
Include also moving resources into their own package
* clusterwaiter
* dnsdomain
* clusterrosaclassic
Include also moving data sources into their own package
* cloudprovider

Rename package provider tp oldprovider

This is done in order to allow changing fwprovider to provider
