+++ oldprovider/common/helpers.go
