+++ oldprovider/rosa_operator_roles_state.go
