+++ provider/cloudprovider/cloud_provider_state.go
@@ -14,12 +14,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package provider
+package cloudprovider
 
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-type DNSDomainState struct {
-	ID types.String `tfsdk:"id"`
+type CloudProviderState struct {
+	ID          string `tfsdk:"id"`
+	Name        string `tfsdk:"name"`
+	DisplayName string `tfsdk:"display_name"`
 }
