+++ provider/cloudprovider/cloud_providers_data_source.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package provider
+package cloudprovider
 
 import (
 	"context"
