+++ provider/clusterrosaclassic/cluster_rosa_classic_resource_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package provider
+package clusterrosaclassic
 
 import (
 	"context"
