+++ provider/clusterwaiter/cluster_waiter_resource.go
@@ -1,4 +1,4 @@
-package provider
+package clusterwaiter
 
 import (
 	"context"
