+++ provider/dnsdomain/dns_domain_resource.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package provider
+package dnsdomain
 
 import (
 	"context"
